requirements.txt
django== 2.0
djangorestframework== 3.7.7